LayoutTests/fast/dom/HTMLTableElement/table-with-invalid-border.html

<!DOCTYPE html>
<head>
<script>
    if (window.testRunner)
        testRunner.dumpAsText();
</script>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
    <p>https://bugs.webkit.org/show_bug.cgi?id=102112: There should be a black box below. <br> 

    <table border="djska" width="100%" height="100%">
        <tr>
            <td height="26"></td>
        </tr>
    </table>
    <div id="console"></div>
    
    <script>
    shouldBeEqualToString('getComputedStyle(document.querySelector("table")).borderTopWidth', "1px");
    </script>
</body>
